putting updated files in correct place in file structure.
