--- admin_seller_request.html
@@ -1,158 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <meta charset="utf-8">
-    <title>Seller Account Requests</title>
-
-    <style>
-        body {
-            font-family: Arial, sans-serif;
-            margin: 0;
-            background-color: #f4f6f8;
-        }
-
-        header {
-            background-color: #2c3e50;
-            color: white;
-            padding: 15px 30px;
-        }
-
-        nav {
-            background-color: #34495e;
-            padding: 10px 30px;
-        }
-
-            nav a {
-                color: white;
-                text-decoration: none;
-                margin-right: 20px;
-                font-weight: bold;
-            }
-
-                nav a:hover {
-                    text-decoration: underline;
-                }
-
-        .container {
-            padding: 30px;
-        }
-
-        table {
-            width: 100%;
-            border-collapse: collapse;
-            background: white;
-            box-shadow: 0 2px 8px rgba(0,0,0,0.1);
-        }
-
-        th, td {
-            padding: 12px;
-            text-align: center;
-        }
-
-        th {
-            background-color: #3498db;
-            color: white;
-        }
-
-        tr:nth-child(even) {
-            background-color: #f2f2f2;
-        }
-
-        .approve-btn {
-            background: #27ae60;
-            color: white;
-            border: none;
-            padding: 6px 10px;
-            margin-right: 5px;
-        }
-
-        .reject-btn {
-            background: #e74c3c;
-            color: white;
-            border: none;
-            padding: 6px 10px;
-        }
-
-        .approve-label {
-            color: rgb(0, 221, 0);
-            font-weight: bold;
-        }
-
-        .rejected-label {
-            color: red;
-            font-weight: bold;
-        }
-
-        .approve-btn:hover {
-            opacity: 0.9;
-        }
-
-        .reject-btn:hover {
-            opacity: 0.9;
-        }
-
-        footer {
-            background-color: #2c3e50;
-            color: white;
-            text-align: center;
-            padding: 15px;
-            margin-top: 40px;
-        }
-    </style>
-</head>
-
-<body>
-
-    <header>
-        <h1>Seller Account Requests</h1>
-    </header>
-
-    <nav>
-        <a href="{% url 'admin_home' %}">Home</a>
-        <a href="{% url 'product_requests' %}">Product Requests</a>
-        <a href="{% url 'seller_requests' %}">Seller Account Requests</a>
-        <a href="{% url 'account_logout' %}">Logout</a>
-    </nav>
-
-    <div class="container">
-
-        <table>
-            <tr>
-                <th>User ID</th>
-                <th>First Name</th>
-                <th>Last Name</th>
-                <th>Email</th>
-                <th>Account Active</th>
-                <th>Response</th>
-            </tr>
-            {% for req in requests %}
-            <tr>
-                <td>{% if req.user %}{{ req.user.id }}{% else %}None{% endif %}</td>
-                <td>{{ req.first_name }}</td>
-                <td>{{ req.last_name }}</td>
-                <td>{{ req.email }}</td>
-                <td>{% if req.user and req.user.is_active %}Yes{% else %}No{% endif %}</td>
-                <td>
-                    {% if req.approved == None %}
-                        <a href="{% url 'approve_seller' req.id %}" class="approve-btn">Approve</a>
-
-                        <a href="{% url 'reject_seller' req.id %}" class="reject-btn">Reject</a>
-
-                    {% elif req.approved == True %}
-                        <span class="approved-label">Approved</span>
-
-                    {% else %}
-                        <span class="rejected-label">Rejected</span>
-                    {% endif %}
-                </td>
-            </tr>
-            {% endfor %}
-        </table>
-    </div>
-
-    <footer>
-        &copy; 2026 ThreadSwitch
-    </footer>
-
-</body>
-</html>
