--- buyer_compare.html
@@ -1,156 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <title>Compare Products</title>
-    <style>
-        body {
-            font-family: Arial, sans-serif;
-            margin: 0;
-            background-color: #f4f6f8;
-            color: #1f2937;
-        }
-
-        header {
-            background-color: #2c3e50;
-            color: white;
-            padding: 15px 30px;
-        }
-
-        nav {
-            background-color: #34495e;
-            padding: 10px 30px;
-        }
-
-        nav a {
-            color: white;
-            text-decoration: none;
-            margin-right: 20px;
-            font-weight: bold;
-        }
-
-        .container {
-            padding: 30px;
-        }
-
-        .compare-grid {
-            display: grid;
-            grid-template-columns: repeat(2, minmax(0, 1fr));
-            gap: 20px;
-        }
-
-        .compare-card {
-            background: white;
-            border-radius: 10px;
-            box-shadow: 0 2px 8px rgba(0, 0, 0, 0.1);
-            padding: 24px;
-        }
-
-        .compare-image,
-        .compare-image-placeholder {
-            width: 100%;
-            height: 260px;
-            border-radius: 10px;
-            margin-bottom: 16px;
-        }
-
-        .compare-image {
-            object-fit: cover;
-            display: block;
-            background: #e5e7eb;
-        }
-
-        .compare-image-placeholder {
-            display: flex;
-            align-items: center;
-            justify-content: center;
-            background: linear-gradient(135deg, #dbeafe, #e5e7eb);
-            color: #475569;
-            font-weight: bold;
-        }
-
-        .compare-row {
-            margin-bottom: 12px;
-        }
-
-        .compare-row strong {
-            display: inline-block;
-            min-width: 95px;
-        }
-
-        .compare-actions {
-            margin-top: 24px;
-            display: flex;
-            gap: 12px;
-            flex-wrap: wrap;
-        }
-
-        .compare-actions a {
-            background: #2563eb;
-            color: white;
-            text-decoration: none;
-            padding: 10px 14px;
-            border-radius: 6px;
-            font-weight: bold;
-        }
-
-        footer {
-            background-color: #2c3e50;
-            color: white;
-            text-align: center;
-            padding: 15px;
-            margin-top: 40px;
-        }
-
-        @media (max-width: 760px) {
-            .compare-grid {
-                grid-template-columns: 1fr;
-            }
-        }
-    </style>
-</head>
-<body>
-    <header>
-        <h1>Compare Products</h1>
-    </header>
-
-    <nav>
-        <a href="{% url 'buyer_home' %}">Home</a>
-        <a href="{% url 'buyer_order_history' %}">Order History</a>
-        <a href="{% url 'cart_view' %}">Cart</a>
-        <a href="{% url 'account_logout' %}">Logout</a>
-    </nav>
-
-    <div class="container">
-        <div class="compare-grid">
-            {% for product in products %}
-                <section class="compare-card">
-                    {% if product.image %}
-                        <img class="compare-image" src="{{ product.image.url }}" alt="{{ product.product_name }}">
-                    {% else %}
-                        <div class="compare-image-placeholder">No Image Available</div>
-                    {% endif %}
-                    <h2>{{ product.product_name }}</h2>
-                    <p class="compare-row"><strong>Category:</strong> {{ product.category }}</p>
-                    <p class="compare-row"><strong>Price:</strong> ${{ product.price }}</p>
-                    <p class="compare-row"><strong>Stock:</strong> {{ product.stock }}</p>
-                    <p class="compare-row"><strong>Seller:</strong> {{ product.seller_id.username }}</p>
-                    <p class="compare-row"><strong>Description:</strong> {{ product.description }}</p>
-                    <div class="compare-actions">
-                        <a href="{% url 'add_to_cart' product.id %}">Add to Cart</a>
-                        <a href="{% url 'remove_from_compare' product.id %}">Remove from Comparison</a>
-                    </div>
-                </section>
-            {% endfor %}
-        </div>
-
-        <div class="compare-actions">
-            <a href="{% url 'buyer_home' %}">Continue Shopping</a>
-            <a href="{% url 'clear_compare' %}" style="background: #dc2626;">Clear Comparison</a>
-        </div>
-    </div>
-
-    <footer>
-        © 2026 ThreadSwitch
-    </footer>
-</body>
-</html>
