+++ buyers/templates/buyer_compare.html
@@ -137,16 +137,15 @@
                     <p class="compare-row"><strong>Description:</strong> {{ product.description }}</p>
                     <div class="compare-actions">
                         <a href="{% url 'add_to_cart' product.id %}">Add to Cart</a>
+                        <a href="{% url 'remove_from_compare' product.id %}">Remove from Comparison</a>
                     </div>
                 </section>
             {% endfor %}
         </div>
 
         <div class="compare-actions">
-            <a href="{{ return_url }}">Back to Catalogue</a>
-            <form method="get" action="{% url 'clear_compare_selections' %}?{% if return_url %}{{ return_url|slice:'1:' }}{% endif %}" style="display: inline;">
-                <button type="submit" style="background: #6b7280; color: white; padding: 10px 14px; border-radius: 6px; font-weight: bold; border: none; cursor: pointer;">New Comparison</button>
-            </form>
+            <a href="{% url 'buyer_home' %}">Continue Shopping</a>
+            <a href="{% url 'clear_compare' %}" style="background: #dc2626;">Clear Comparison</a>
         </div>
     </div>
